Move cli functionality from main.rs into cli.rs
