+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod algorithms;
+pub mod cli;
 pub mod commands;
 pub mod factorization;
 pub mod orchestration;
